move the cli/debug package to cmd/dockerd/debug
